[R] test: move test from CarRacing to Cars that is more core level
